Moved seo-store tests to separate folder and added coverage report
